Allow external Guzzle Response injection
As it was posible in the JsonContext to inject external JSON accessing the Inspector to third party Context, was impossible to inject a Response when for to obtain it is more complex than a simple array of strings    
(Ex. multipart with files) , even having access to RestApiBrowser. 
This small change delegates the "breaking" of the Rest rules , than time to time in real world APIs happens, to third party Context but still allow to use proper check code and the json validations steps without need to reimplement them.